Character search, leaderboard changed slightly, a way to gain rating and achievement ranking added.

+++ MiniArmory.Core/Models/LeaderboardViewModel.cs
@@ -6,7 +6,7 @@ namespace MiniArmory.Core.Models
     {
         public Guid Id { get; set; }
 
-        public short Rating { get; set; }
+        public int Rating { get; set; }
 
         public string Name { get; set; }
